refactor: fetch를 이전 단계에서 수행
